+++ Src/interpolators.c
@@ -85,7 +85,7 @@ uint8_t getLampIndex(Tasks tasks,uint8_t lampnr)
 	return 0xFF;
 }
 
-uint8_t setColorFramesInterpolation(Tasks tasks, uint8_t r,uint8_t g,uint8_t b,int16_t frames,uint8_t interpolation, uint8_t lamp_nr,uint8_t step)
+uint8_t setColorFramesInterpolation(Tasks tasks, uint8_t r,uint8_t g,uint8_t b,int32_t frames,uint8_t interpolation, uint8_t lamp_nr,uint8_t step)
 {
 	uint8_t idx;
 	idx=getLampIndex(tasks,lamp_nr);
